+++ web/src/echo.js
@@ -10,6 +10,7 @@ import * as local_message from "./local_message";
 import * as markdown from "./markdown";
 import * as message_events from "./message_events";
 import * as message_lists from "./message_lists";
+import * as message_live_update from "./message_live_update";
 import * as message_store from "./message_store";
 import * as narrow_state from "./narrow_state";
 import * as notifications from "./notifications";
@@ -23,7 +24,6 @@ import * as sent_messages from "./sent_messages";
 import * as stream_list from "./stream_list";
 import * as stream_topic_history from "./stream_topic_history";
 import * as transmit from "./transmit";
-import * as ui from "./ui";
 import * as util from "./util";
 
 // Docs: https://zulip.readthedocs.io/en/latest/subsystems/sending-messages.html
@@ -61,9 +61,25 @@ function insert_message(message) {
     message_events.insert_new_messages([message], true);
 }
 
+function show_message_failed(message_id, failed_msg) {
+    // Failed to send message, so display inline retry/cancel
+    message_live_update.update_message_in_all_views(message_id, ($row) => {
+        const $failed_div = $row.find(".message_failed");
+        $failed_div.toggleClass("hide", false);
+        $failed_div.find(".failed_text").attr("title", failed_msg);
+    });
+}
+
+function show_failed_message_success(message_id) {
+    // Previously failed message succeeded
+    message_live_update.update_message_in_all_views(message_id, ($row) => {
+        $row.find(".message_failed").toggleClass("hide", true);
+    });
+}
+
 function failed_message_success(message_id) {
     message_store.get(message_id).failed_request = false;
-    ui.show_failed_message_success(message_id);
+    show_failed_message_success(message_id);
 }
 
 function resend_message(message, $row) {
@@ -442,7 +458,7 @@ export function _patch_waiting_for_ack(data) {
 export function message_send_error(message_id, error_response) {
     // Error sending message, show inline
     message_store.get(message_id).failed_request = true;
-    ui.show_message_failed(message_id, error_response);
+    show_message_failed(message_id, error_response);
 }
 
 function abort_message(message) {
