+++ web/src/ui.js
@@ -1,8 +1,6 @@
 import $ from "jquery";
 import SimpleBar from "simplebar";
 
-import * as message_lists from "./message_lists";
-
 // What, if anything, obscures the home tab?
 
 export function replace_emoji_with_text($element) {
@@ -44,56 +42,6 @@ export function reset_scrollbar($element) {
     } else {
         element.scrollTop = 0;
     }
-}
-
-function update_message_in_all_views(message_id, callback) {
-    for (const msg_list of message_lists.all_rendered_message_lists()) {
-        const $row = msg_list.get_row(message_id);
-        if ($row === undefined) {
-            // The row may not exist, e.g. if you do an action on a message in
-            // a narrowed view
-            continue;
-        }
-        callback($row);
-    }
-}
-
-export function update_starred_view(message_id, new_value) {
-    const starred = new_value;
-
-    // Avoid a full re-render, but update the star in each message
-    // table in which it is visible.
-    update_message_in_all_views(message_id, ($row) => {
-        const $elt = $row.find(".star");
-        const $star_container = $row.find(".star_container");
-        if (starred) {
-            $elt.addClass("fa-star").removeClass("fa-star-o");
-            $star_container.removeClass("empty-star");
-        } else {
-            $elt.removeClass("fa-star").addClass("fa-star-o");
-            $star_container.addClass("empty-star");
-        }
-        const data_template_id = starred
-            ? "unstar-message-tooltip-template"
-            : "star-message-tooltip-template";
-        $star_container.attr("data-tooltip-template-id", data_template_id);
-    });
-}
-
-export function show_message_failed(message_id, failed_msg) {
-    // Failed to send message, so display inline retry/cancel
-    update_message_in_all_views(message_id, ($row) => {
-        const $failed_div = $row.find(".message_failed");
-        $failed_div.toggleClass("hide", false);
-        $failed_div.find(".failed_text").attr("title", failed_msg);
-    });
-}
-
-export function show_failed_message_success(message_id) {
-    // Previously failed message succeeded
-    update_message_in_all_views(message_id, ($row) => {
-        $row.find(".message_failed").toggleClass("hide", true);
-    });
 }
 
 // Save the compose content cursor position and restore when we
